Want the y-position to be an f64 when running the pixel program (most algorithms will use a floating point value here, this is helpful for implementing super-sampling)

+++ render_software/src/pixel/pixel_program.rs
@@ -18,7 +18,7 @@ pub trait PixelProgram : Send {
     ///
     /// The target points to the start of the range of values to be written. `x_range` provides the range of X values to fill with pixels.
     ///
-    fn draw_pixels(&self, pixel_program_cache: &PixelProgramCache<Self::Pixel>, data_cache: &PixelProgramDataCache<Self::Pixel>, target: &mut [Self::Pixel], x_range: Range<i32>, y_pos: i32, program_data: &Self::ProgramData);
+    fn draw_pixels(&self, pixel_program_cache: &PixelProgramCache<Self::Pixel>, data_cache: &PixelProgramDataCache<Self::Pixel>, target: &mut [Self::Pixel], x_range: Range<i32>, y_pos: f64, program_data: &Self::ProgramData);
 }
 
 ///
@@ -28,7 +28,7 @@ pub trait PixelProgram : Send {
 ///
 pub struct PixelProgramFn<TFn, TPixel, TData>
 where 
-    TFn: Send + Fn(&mut [TPixel], Range<i32>, i32, &TData) -> (),
+    TFn: Send + Fn(&mut [TPixel], Range<i32>, f64, &TData) -> (),
 {
     /// The function to call to fill in the pixels
     function: TFn,
@@ -44,7 +44,7 @@ where
 ///
 pub struct PerPixelProgramFn<TFn, TPixel, TData>
 where 
-    TFn: Fn(i32, i32, &TData) -> TPixel,
+    TFn: Fn(i32, f64, &TData) -> TPixel,
 {
     /// The function to call to fill in the pixels
     function: TFn,
@@ -55,7 +55,7 @@ where
 
 impl<TFn, TPixel, TData> From<TFn> for PixelProgramFn<TFn, TPixel, TData> 
 where 
-    TFn: Send + Fn(&mut [TPixel], Range<i32>, i32, &TData) -> (),
+    TFn: Send + Fn(&mut [TPixel], Range<i32>, f64, &TData) -> (),
 {
     fn from(function: TFn) -> Self {
         PixelProgramFn {
@@ -67,7 +67,7 @@ where
 
 impl<TFn, TPixel, TData> PixelProgram for PixelProgramFn<TFn, TPixel, TData> 
 where 
-    TFn:    Send + Fn(&mut [TPixel], Range<i32>, i32, &TData) -> (),
+    TFn:    Send + Fn(&mut [TPixel], Range<i32>, f64, &TData) -> (),
     TData:  Send,
     TPixel: Send,
 {
@@ -75,14 +75,14 @@ where
     type ProgramData    = TData;
 
     #[inline]
-    fn draw_pixels(&self, _: &PixelProgramCache<Self::Pixel>, _: &PixelProgramDataCache<Self::Pixel>, target: &mut [TPixel], x_range: Range<i32>, ypos: i32, program_data: &TData) {
+    fn draw_pixels(&self, _: &PixelProgramCache<Self::Pixel>, _: &PixelProgramDataCache<Self::Pixel>, target: &mut [TPixel], x_range: Range<i32>, ypos: f64, program_data: &TData) {
         (self.function)(target, x_range, ypos, program_data)
     }
 }
 
 impl<TFn, TPixel, TData> From<TFn> for PerPixelProgramFn<TFn, TPixel, TData> 
 where 
-    TFn: Fn(i32, i32, &TData) -> TPixel,
+    TFn: Fn(i32, f64, &TData) -> TPixel,
 {
     fn from(function: TFn) -> Self {
         PerPixelProgramFn {
@@ -94,7 +94,7 @@ where
 
 impl<TFn, TPixel, TData> PixelProgram for PerPixelProgramFn<TFn, TPixel, TData> 
 where 
-    TFn:    Send + Fn(i32, i32, &TData) -> TPixel,
+    TFn:    Send + Fn(i32, f64, &TData) -> TPixel,
     TData:  Send,
     TPixel: Send,
 {
@@ -102,7 +102,7 @@ where
     type ProgramData    = TData;
 
     #[inline]
-    fn draw_pixels(&self, _: &PixelProgramCache<Self::Pixel>, _: &PixelProgramDataCache<Self::Pixel>, target: &mut [TPixel], x_range: Range<i32>, ypos: i32, program_data: &TData) {
+    fn draw_pixels(&self, _: &PixelProgramCache<Self::Pixel>, _: &PixelProgramDataCache<Self::Pixel>, target: &mut [TPixel], x_range: Range<i32>, ypos: f64, program_data: &TData) {
         let mut pos = 0;
         for x in x_range {
             target[pos] = (self.function)(x, ypos, program_data);
